Implement System class as a singleton
This will allow you to easily shoot yourself in the foot, but also allows for objects that need
to know about each other to actually know about each other. Notably Fibers know what Body/site
they're on, but not where that Body is or what to do with it. Rather than using a pointer to a
body (which needs to be found, ensured is accurate, and requires forward declarations and other
things), the user can request the Body directly (through appropriate getters!).